requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.8.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
databases==0.6.1
frozenlist==1.3.1
greenlet==1.1.3.post0
idna==3.4
Mako==1.2.3
MarkupSafe==2.1.1
multidict==6.0.2
psycopg2-binary==2.9.4
python-dotenv==0.21.0
pytz==2022.5
SQLAlchemy==1.4.42
yarl==1.8.1
